+++ src/WeihanLi.EntityFramework/EntityFrameworkExtensions.cs
@@ -1,6 +1,0 @@
-﻿namespace WeihanLi.EntityFramework
-{
-    public static class EntityFrameworkExtensions
-    {
-    }
-}