+++ src/WeihanLi.EntityFramework/IEFRepositoryFactory.cs
@@ -2,7 +2,7 @@
 
 namespace WeihanLi.EntityFramework
 {
-    public interface IEFRepositoryFactory<TDbContext> where TDbContext : DbContext
+    public interface IEFRepositoryFactory<out TDbContext> where TDbContext : DbContext
     {
         IEFRepository<TDbContext, TEntity> GetRepository<TEntity>() where TEntity : class;
     }