+++ test/WeihanLi.EntityFramework.Test/EFTestBase.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using Microsoft.Extensions.DependencyInjection;
+using Xunit;
+
+namespace WeihanLi.EntityFramework.Test
+{
+    public class EFTestBase : IClassFixture<EFTestFixture>
+    {
+        public IServiceProvider Services { get; }
+        public IEFRepository<TestDbContext, TestEntity> Repository { get; }
+
+        public EFTestBase(EFTestFixture fixture)
+        {
+            Services = fixture.Services;
+            Repository = fixture.Services
+                .GetRequiredService<IEFRepository<TestDbContext, TestEntity>>();
+            Repository.DbContext.CleanData();
+        }
+    }
+}